+++ src/MSALWrapper/PCAWrapper.cs
@@ -42,12 +42,10 @@ namespace Microsoft.Authentication.MSALWrapper
         /// <param name="pca">The public client application instance.</param>
         /// <param name="errors">The errors list to append error encountered to.</param>
         /// <param name="tenantId">The tenant ID to help key the cache off of.</param>
-        /// <param name="osxKeyChainSuffix">An optional (can be null) suffix to further customize key chain token caches on OSX.</param>
-        /// <param name="cacheFilePath">The cache filename.</param>
-        public PCAWrapper(ILogger logger, IPublicClientApplication pca, IList<Exception> errors, Guid tenantId, string osxKeyChainSuffix, string cacheFilePath)
+        public PCAWrapper(ILogger logger, IPublicClientApplication pca, IList<Exception> errors, Guid tenantId)
             : this(logger, pca)
         {
-            var cacher = new PCACache(logger, tenantId, cacheFilePath, osxKeyChainSuffix);
+            var cacher = new PCACache(logger, tenantId);
             cacher.SetupTokenCache(this.pca.UserTokenCache, errors);
         }